pyproject.toml
[project]
name = "ed-bot"
version = "0.1.0"
description = "EdStem forum automation. Ingests Q&A history, generates draft answers with guardrails."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "ed-api",
    "pyqmd",
    "markitdown>=0.1.0",
]

[project.optional-dependencies]
lectures = [
    "faster-whisper>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
ed = "ed_bot.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ed_bot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv.sources]
ed-api = { path = "../ed-api", editable = true }
pyqmd = { path = "../py-qmd", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
